Make precision factor a double instead of an int Use l2Norm when adjusting the error margin to get more accurate results

+++ src/main/java/org/moltimate/moltimatebackend/dto/ActiveSiteAlignmentRequest.java
@@ -31,7 +31,7 @@ public class ActiveSiteAlignmentRequest {
     private List<String> filters = new ArrayList<>();
     private List<MultipartFile> customMotifs = new ArrayList<>();
     private String ecNumber;
-    private int precisionFactor;
+    private double precisionFactor;
 
     public PdbQueryResponse callPdbForResponse() {
         PdbQueryResponse response = ProteinUtils.queryPdbResponse(pdbIds);
@@ -60,9 +60,9 @@ public class ActiveSiteAlignmentRequest {
         return results;
     }
 
-    public int getPrecisionFactor() {
+    public double getPrecisionFactor() {
         if (this.precisionFactor <= 0) {
-            return 1;
+            return 1d;
         }
         return this.precisionFactor;
     }
